Updated restaurant classes with variable type changes - owneruser owner - ArrayList<Review> reviews - errors unfixed: string identifiers to and from objects (reviews, owner)

+++ src/main/java/restaurant_use_case/RestaurantDSRequestModel.java
@@ -1,6 +1,7 @@
 package restaurant_use_case;
 
 import entities.OwnerUser;
+import entities.Review;
 
 import java.util.ArrayList;
 
@@ -10,14 +11,14 @@ public class RestaurantDSRequestModel {
     private String cuisineType;
     private int priceBucket;
     private String owner;
-    private int avgStars;
+    private double avgStars;
     private ArrayList<String> reviews;
 
     public RestaurantDSRequestModel(String owner, String name, String location,
                                   String cuisineType, int priceBucket) {
         this.name = name;
         this.location = location;
-        this.owner = owner;
+        this.owner = owner; // convert to string
         this.cuisineType = cuisineType;
         this.priceBucket= priceBucket;
         this.avgStars = 0;
@@ -25,15 +26,15 @@ public class RestaurantDSRequestModel {
     }
 
     public RestaurantDSRequestModel(String owner, String name, String location,
-                                    String cuisineType, int priceBucket, int avgStars,
+                                    String cuisineType, int priceBucket, double avgStars,
                                     ArrayList<String> reviews) {
         this.name = name;
         this.location = location;
-        this.owner = owner;
+        this.owner = owner; // convert to string identifier
         this.cuisineType = cuisineType;
         this.priceBucket= priceBucket;
         this.avgStars = avgStars;
-        this.reviews = reviews;
+        this.reviews = reviews; // convert to strings
     }
 
     // Retrieve Information
@@ -42,6 +43,6 @@ public class RestaurantDSRequestModel {
     String getCuisineType() {return this.cuisineType;}
     int getPriceBucket() {return this.priceBucket;}
     String getOwner() {return this.owner;}
-    int getAvgStars() {return this.avgStars;}
+    double getAvgStars() {return this.avgStars;}
     ArrayList<String> getReviews() {return this.reviews;}
 }
